Fixes #45 -- AttributeError while resolving the user model in Django
Under some conditions ``UserSocialAuth._meta.get_field('user').rel.to`` does not return a Model class but a string of the form ``'<app_label>.<model_name>'``. This results in an AttributeError in the Django admin.py which ultimately triggers an error in Django's RegexURLResolver. This is fixed by properly resolving the user model and return a class rather than a string.